#4096: Fix issue with DPRINT server closing too early for some WAITs
Issue here was that for some tests with lots of RAISE/WAITs, the DPRINT
server would sometimes close before capturing all prints. Solution here
is to adjust the exit condition to account for existing WAITs. This
works fine as long as all WAITs are valid, and handling for any invalid
WAITs is handled in issue #4073.